// sdk/typescript/src/client/generated/transport_kernel.ts
export interface RawRestResponse {
  readonly status: number;
  readonly headers: readonly (readonly [string, string])[];
  readonly body: Uint8Array;
}